fix(minio):fix the issue if missing ingress
- refactor the reconcile process
- remove filed variables from the minio controller
- rename pkg files
- remove useless pkg files


- fix #475
- fix #433
