+++ pkg/cluster/controllers/storage/constants.go
@@ -2,15 +2,14 @@ package storage
 
 // Define error message.
 const (
-	GetMinIOError            = "Get minIO error"
-	GenerateMinIOCrError     = "Generate minIO cr error"
-	UpdateMinIOError         = "Update minIO error"
-	CreateMinIOSecretError   = "Create minIO secret error" // nolint:gosec
-	CreateMinIOServiceError  = "Create service of minIO error"
-	CreateMinIOIngressError  = "Create ingress of minIO error"
-	CreateMinIOError         = "Create minIO CR error"
-	ScaleMinIOError          = "Scale minIO error"
-	CreateDefaultBucketError = "Create default bucket in minIO Error"
-	getMinIOProperties       = "Get MinIO Properties error"
-	updateIngressError       = "update minIO ingress error"
+	GetMinIOError            = "get minIO error"
+	GenerateMinIOCrError     = "generate minIO cr error"
+	UpdateMinIOError         = "update minIO error"
+	CreateMinIOSecretError   = "create minIO secret error" // nolint:gosec
+	CreateMinIOIngressError  = "create ingress of minIO error"
+	CreateMinIOError         = "create minIO CR error"
+	CreateDefaultBucketError = "create default bucket in minIO Error"
+	GetMinIOProperties       = "get MinIO Properties error"
+	UpdateIngressError       = "update minIO ingress error"
+	GetMinIOIngressError     = "get minIO ingress error"
 )
